Reorganized URLs and split source code for different interfaces.
Move '/cat' to '/opds', to leave room for '/cat' to hold the HTML
interface, and separated the OPDS interface code into a new file,
to avoid namespace problems, since the HTML and the OPDS
interfaces will both have 'author' functions in them, etc. Also
created "catutils" file for code that all of the admin, HTML, and
OPDS interfaces will need.